% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logger.R
\name{get_dhsc_logger}
\alias{get_dhsc_logger}
\title{Get a log collection}
\usage{
get_dhsc_logger(log_file = "output/log.txt", prefix = "log")
}
\arguments{
\item{log_file}{Path to log file (will be created if not present)}

\item{prefix}{Prefix to use on log names}
}
\value{
A log collection
}
\description{
Return a log collection with one log to write to
console and another to write to file. Logs are named
\code{prefix}.console and \code{prefix}.file.
}
