% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logger.R
\name{LogCollection}
\alias{LogCollection}
\title{R6 Class wrapping a collection of \code{futile.logger} logs}
\description{
Wrapper for a collection of \code{futile.logger} logs. Allows each log
to be given a separate threshold but only needs a single call to
write to them all. Order of logging levels is (see \code{futile.logger}
for more details):
[ \code{TRACE} | \code{DEBUG} | \code{INFO} | \code{WARN} | \code{ERROR} | \code{FATAL} ]
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-LogCollection-trace}{\code{LogCollection$trace()}}
\item \href{#method-LogCollection-debug}{\code{LogCollection$debug()}}
\item \href{#method-LogCollection-info}{\code{LogCollection$info()}}
\item \href{#method-LogCollection-warn}{\code{LogCollection$warn()}}
\item \href{#method-LogCollection-error}{\code{LogCollection$error()}}
\item \href{#method-LogCollection-fatal}{\code{LogCollection$fatal()}}
\item \href{#method-LogCollection-add}{\code{LogCollection$add()}}
\item \href{#method-LogCollection-set_threshold}{\code{LogCollection$set_threshold()}}
\item \href{#method-LogCollection-get_log_names}{\code{LogCollection$get_log_names()}}
\item \href{#method-LogCollection-clone}{\code{LogCollection$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LogCollection-trace"></a>}}
\if{latex}{\out{\hypertarget{method-LogCollection-trace}{}}}
\subsection{Method \code{trace()}}{
Write to log at \code{TRACE} level.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LogCollection$trace(msg, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{msg}}{(\code{character()})\cr
The message to log (can be \code{sprintf} format string).}

\item{\code{...}}{\cr
Optional arguments to populate the msg format string.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LogCollection-debug"></a>}}
\if{latex}{\out{\hypertarget{method-LogCollection-debug}{}}}
\subsection{Method \code{debug()}}{
Write to log at \code{DEBUG} level.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LogCollection$debug(msg, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{msg}}{(\code{character()})\cr
The message to log (can be \code{sprintf} format string).}

\item{\code{...}}{\cr
Optional arguments to populate the msg format string.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LogCollection-info"></a>}}
\if{latex}{\out{\hypertarget{method-LogCollection-info}{}}}
\subsection{Method \code{info()}}{
Write to log at \code{INFO} level.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LogCollection$info(msg, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{msg}}{(\code{character()})\cr
The message to log (can be \code{sprintf} format string).}

\item{\code{...}}{\cr
Optional arguments to populate the msg format string.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LogCollection-warn"></a>}}
\if{latex}{\out{\hypertarget{method-LogCollection-warn}{}}}
\subsection{Method \code{warn()}}{
Write to log at \code{WARN} level.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LogCollection$warn(msg, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{msg}}{(\code{character()})\cr
The message to log (can be \code{sprintf} format string).}

\item{\code{...}}{\cr
Optional arguments to populate the msg format string.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LogCollection-error"></a>}}
\if{latex}{\out{\hypertarget{method-LogCollection-error}{}}}
\subsection{Method \code{error()}}{
Write to log at \code{ERROR} level.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LogCollection$error(msg, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{msg}}{(\code{character()})\cr
The message to log (can be \code{sprintf} format string).}

\item{\code{...}}{\cr
Optional arguments to populate the msg format string.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LogCollection-fatal"></a>}}
\if{latex}{\out{\hypertarget{method-LogCollection-fatal}{}}}
\subsection{Method \code{fatal()}}{
Write to log at \code{FATAL} level.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LogCollection$fatal(msg, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{msg}}{(\code{character()})\cr
The message to log (can be \code{sprintf} format string).}

\item{\code{...}}{\cr
Optional arguments to populate the msg format string.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LogCollection-add"></a>}}
\if{latex}{\out{\hypertarget{method-LogCollection-add}{}}}
\subsection{Method \code{add()}}{
Add a log to the logger collection.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LogCollection$add(log_name, log_appender = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{log_name}}{(\code{character()})\cr
Name of log to add to logger collection.}

\item{\code{log_appender}}{(\code{futile.logger} appender)\cr
Appender type see \code{futile.logger} for details.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LogCollection-set_threshold"></a>}}
\if{latex}{\out{\hypertarget{method-LogCollection-set_threshold}{}}}
\subsection{Method \code{set_threshold()}}{
Set threshold of log in logger collection.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LogCollection$set_threshold(log_name, log_threshold)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{log_name}}{(\code{character()}) \cr
Name of log to add to logger.}

\item{\code{log_threshold}}{("TRACE", "DEBUG", "INFO", "WARN", "ERROR", "FATAL")\cr
Level to set threshold at.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LogCollection-get_log_names"></a>}}
\if{latex}{\out{\hypertarget{method-LogCollection-get_log_names}{}}}
\subsection{Method \code{get_log_names()}}{
Get all logs in logger collection.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LogCollection$get_log_names()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LogCollection-clone"></a>}}
\if{latex}{\out{\hypertarget{method-LogCollection-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LogCollection$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
